Update DUC scripts.

+++ process_duc.py
@@ -1,10 +1,8 @@
-from __future__ import print_function
-
-import NeuralSum as ns
-
+import fire
 import logging as log
 import os
-import sys
+
+import NeuralSum as ns
 
 """
 script for processing DUC datasets (2003 or 2004) for use in evaluation
@@ -13,18 +11,26 @@ script for processing DUC datasets (2003 or 2004) for use in evaluation
     - sentences.txt for the first sentence of each article.
     - summaries.txt for the target summary.
 - line n of sentences.txt corresponds to line n of summaries.txt
-- since there are about 4 summaries per sentence, there will be repeats in
+- since there are 4 summaries per sentence, there will be repeats in
     sentences.txt.
 - utilizes NeuralSum/preprocessing.
+- defaults to processing DUC2004 if no argument is provided.
 - run options:
-    >>> python process_duc.py 2003
-    >>> python process_duc.py 2004
+    >>> python process_duc.py [--which_duc=(2003|2004)]
 
 make sure to rerun this script whenever preprocessing steps have been changed!
 deletion of sentences.txt and summaries.txt is done automatically.
 """
 
-def main(sentences, summaries):
+def main(which_duc=2004):
+    if type(which_duc) != int or which_duc not in [2003, 2004]:
+        msg = 'which_duc must be either 2003 or 2004.'
+        log.exception(msg)
+        raise ValueError(msg)
+
+    sentences = './data/duc' + str(which_duc) + '/sentences.txt'
+    summaries = './data/duc' + str(which_duc) + '/summaries.txt'
+
     # remove existing files:
     try:
         os.remove(sentences)
@@ -38,27 +44,20 @@ def main(sentences, summaries):
         pass
 
     # preprocess articles:
-    if '2004' in sentences:
-        log.info("Loading DUC2004 articles")
-        duc_articles = ns.parse_duc_2004()
-    else:
-        log.info("Loading DUC2003 articles")
-        duc_articles = ns.parse_duc_2003()
-
-    sentences_gen = sentence_generator(duc_articles)
-    summaries_gen = summary_generator(duc_articles)
+    log.info('Loading DUC' + str(which_duc) + ' articles')
+    duc_articles = ns.parse_duc_2004() if which_duc == 2004 else ns.parse_duc_2003()
 
     # write sentences and summaries to repspective files:
     sent_count = 0
     with open(sentences, mode='a') as sen_f:
-        for _, sen in enumerate(sentences_gen):
+        for _, sen in enumerate(sentence_generator(duc_articles)):
             sen_f.write(sen + '\n')
             sent_count += 1
         log.info("Wrote " + str(sent_count) + " sentences to: " + sentences)
 
     summ_count = 0
     with open(summaries, mode='a') as summ_f:
-        for _, summ in enumerate(summaries_gen):
+        for _, summ in enumerate(summary_generator(duc_articles)):
             summ_f.write(summ + '\n')
             summ_count += 1
         log.info("Wrote " + str(summ_count) + " summaries to: " + summaries)
@@ -77,17 +76,4 @@ def summary_generator(articles):
 
 if __name__ == '__main__':
     log.getLogger()
-    if len(sys.argv) > 3:
-        raise ValueError("Argument count too high.")
-    if len(sys.argv) == 2:
-        if sys.argv[1] == '2004':
-            sentences = './data/duc2004/sentences.txt'
-            summaries = './data/duc2004/summaries.txt'
-        elif sys.argv[1] == '2003':
-            sentences = './data/duc2003/sentences.txt'
-            summaries = './data/duc2003/summaries.txt'
-        else:
-            raise ValueError("Argument must be either '2004' or '2003'")
-        main(sentences, summaries)
-    else:
-        main('./data/duc2004/sentences.txt','./data/duc2004/summaries.txt')
+    fire.Fire(main)
